feat(metrics): entropia_linha e entropia_coluna por concurso (bits)
Implementa entropia_linha_por_concurso@1.0.0 e entropia_coluna_por_concurso@1.0.0
conforme catálogo (Shannon em bits sobre a distribuição por linha/coluna).

WindowMetricValue passa a expor valores numéricos como double para séries em
bits; métricas de contagem mantêm inteiros como doubles. Registro no
dispatcher, validador V0, tool HTTP/MCP e testes determinísticos.

Made-with: Cursor

+++ src/LotofacilMcp.Application/UseCases/UseCaseContracts.cs
@@ -23,7 +23,7 @@ public sealed record FrequencyMetricValueView(
     string Unit,
     string Version,
     WindowDescriptor Window,
-    IReadOnlyList<int> Value,
+    IReadOnlyList<double> Value,
     string Explanation);
 
 public sealed record StabilityRankingEntryView(
